backend (1)/backend/requirements.txt
fastapi
uvicorn
pydantic
supabase
httpx
python-dotenv
